requirements.txt
flask
flask-cors
numpy
pandas
joblib
scikit-learn
xgboost
gunicorn
supabase
python-dotenv
